requirements.txt
# Core dependencies for Library Management System Database

# Database driver
psycopg2-binary>=2.9.0  # PostgreSQL driver

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Utilities
python-dotenv>=1.0.0  # Environment variable management
click>=8.0.0  # CLI utilities

# Optional: For advanced features
# sqlalchemy>=2.0.0  # ORM (if needed later)
# alembic>=1.8.0  # Alternative migration tool